requirements.txt
Django==2.2.3
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-richtextfield==1.6.1
fontawesome-free==5.15.3
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
